Move from fastapi to quart
There are some loose ends still, but the migration should be
functionally complete. The primary motivation for this move was to be
able have hot server reloading in dev mode, and that works.

Regression: we have to run the API server and frontend dev server
separately right now -- otherwise, the dev server respawns every time
the API server reloads. Will look into ways of fixing this.